Rename writers directory to handlers
