+++ handlers/__init__.py
@@ -1,4 +1,4 @@
-from writers.config_manager import ConfigManager
+from handlers.config_manager import ConfigManager
 
 # Create a global instance of the ConfigManager class
 config_manager = ConfigManager()
